use Contexts for message selection
